# openapi.yml
...
openapi: 3.0.3
info:
  title: approvald
  description: |
    One-tap approval gating. A machine registers a nonce and gets two
    single-use URLs; a person taps one on their phone; the decision is
    published to ntfy for the waiting machine to read.
  version: 0.1.0
paths:
...
  /d/{nonce}/{token}/{verb}:
    parameters:
      - { $ref: '#/components/parameters/Nonce' }
      - { $ref: '#/components/parameters/Token' }
      - { $ref: '#/components/parameters/Verb' }
    get:
      operationId: tapConfirm
      summary: The confirmation page a person lands on
      description: |
        HTML, because the consumer is a thumb on a phone. Renders a
        button that POSTs to this same path - a GET must not decide
        anything, since link previews and prefetchers issue GETs.
      responses:
        default:
          description: An error this operation does not otherwise document.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '200':
          description: A page with a button confirming the decision.
          content:
            text/html:
              schema: { type: string }
        '404':
          description: An explanation for a person.
          content:
            text/html:
              schema: { type: string }
        '410':
          description: An explanation for a person.
          content:
            text/html:
              schema: { type: string }
    post:
      operationId: tap
      summary: Record the decision and publish it
      responses:
        default:
          description: An error this operation does not otherwise document.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '200':
          description: The decision was recorded and published.
          content:
            text/html:
              schema: { type: string }
        '404':
          description: An explanation for a person.
          content:
            text/html:
              schema: { type: string }
        '410':
          description: An explanation for a person.
          content:
            text/html:
              schema: { type: string }
        '502':
          description: An explanation for a person.
          content:
            text/html:
              schema: { type: string }

  # Any operation may fail in a way it does not document. `default` is
  # what lets ogen generate an ErrorStatusCode for those, so a handler
  # error becomes a documented JSON body rather than an empty 500.
...
components:
  securitySchemes:
    registerToken:
      type: http
      scheme: bearer
      description: The shared secret in REGISTER_TOKEN.
  parameters:
    Nonce:
      name: nonce
      in: path
      required: true
      schema: { type: string, maxLength: 128 }
    Token:
      name: token
      in: path
      required: true
      description: |
        A single-use capability. It is in the URL, so nothing may log or
        echo the request path - see the note in server.go.
      schema: { type: string }
    Verb:
      name: verb
      in: path
      required: true
      schema: { type: string, enum: [approve, deny] }
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }